requirements-dev.txt
# pylint
astroid==1.6.5
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.2
wrapt==1.10.11

# flake8
flake8==3.5.0
pycodestyle==2.3.1
pyflakes==1.6.0

# pytest
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.2
pytest-flask==0.10.0

-r requirements.txt